# gathering-service/api/src/main/resources/api.yaml
# API Template to extend when a service is created based on this archetype.
# Do not create a new .yml file next to this, rather use this as the main descriptor.

openapi: "3.0.3"
info:
  version: 1.0.0
  title: Microservice

servers:
  - url: http://gathering-service:8081

paths:

  # ==============================================================================
  # Gathering
  # ==============================================================================

  /api/v1/gatherings:
    get:
      tags:
        - gathering
      operationId: getGatherings
      description: Get every gathering
      responses:
        200:
          description: List of gatherings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GatheringDto'
    post:
      tags:
        - gathering
      operationId: createGathering
      description: Create a Gathering
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/GatheringDto'
        description: Content of the Gathering
        required: true
      responses:
        201:
          description: The created Gathering
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatheringDto'

  /api/v1/gatherings/{ownerId}:
    get:
      tags:
        - gathering
      operationId: getGatheringsByOwnerId
      description: Get every gathering by specified ownerId
      parameters:
        - name: ownerId
          in: path
          required: true
          description: Id of the owner
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: List of gatherings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GatheringDto'


components:
  schemas:

    # ==============================================================================
    # Enums
    # ==============================================================================

    # ==============================================================================
    # Dtos
    # ==============================================================================

    GatheringDto:
      required:
        - ownerId
      properties:
        id:
          type: integer
          format: int64
        ownerId:
          type: integer
          format: int64
          minimum: 1
        description:
          type: string